Reading in GeoTiffs is now done through their crop method
Updates tests

Removed old Spec Name
